bin: Extract business logic into `admin` modules
