+++ src/admin/mod.rs
@@ -1,2 +1,9 @@
+pub mod delete_crate;
+pub mod delete_version;
 pub mod dialoguer;
 pub mod on_call;
+pub mod populate;
+pub mod render_readmes;
+pub mod test_pagerduty;
+pub mod transfer_crates;
+pub mod verify_token;
